% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioThermR-package.R
\docType{package}
\name{BioThermR-package}
\alias{BioThermR-package}
\title{BioThermR: Standardized Processing and Analysis of Thermal Imaging Data in Animal Studies}
\description{
A modular framework for standardized analysis of thermal imaging data in animal experimentation. The package integrates thermographic data import (FLIR, raw, CSV), automated region of interest (ROI) segmentation based on 'EBImage' (Pau et al., 2010 \doi{10.1093/bioinformatics/btq046}), interactive ROI refinement, and high-throughput batch processing.

A modular framework for standardized analysis of thermal imaging data in animal experimentation. The package integrates thermographic data import (FLIR, raw, CSV), automated region of interest (ROI) segmentation based on 'EBImage' (Pau et al., 2010 \doi{10.1093/bioinformatics/btq046}), interactive ROI refinement, and high-throughput batch processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RightSZ/BioThermR}
  \item \url{https://rightsz.github.io/BioThermR/}
  \item Report bugs at \url{https://github.com/RightSZ/BioThermR/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/RightSZ/BioThermR}
  \item \url{https://rightsz.github.io/BioThermR/}
  \item Report bugs at \url{https://github.com/RightSZ/BioThermR/issues}
}

}
\author{
\strong{Maintainer}: BeiHao Li \email{szright2000@gmail.com}

}
\keyword{internal}
